Tweaked CLI to encompass creating setting if not there, renaming
